requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pytrends==4.9.2
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.10.11
